Picker fixes

+++ vistrails/packages/CPCViewer/DistributedPointCollections.py
@@ -314,7 +314,7 @@ class vtkPointCloud(QtCore.QObject):
             else: return
         vtk_points_data = numpy_support.numpy_to_vtk( self.np_points_data )    
         vtk_points_data.SetNumberOfComponents( 3 )
-        vtk_points_data.SetNumberOfTuples( len( self.np_points_data ) / 3 )     
+        vtk_points_data.SetNumberOfTuples( int( len( self.np_points_data ) / 3 ) )     
         self.vtk_planar_points = vtk.vtkPoints()
         self.vtk_planar_points.SetData( vtk_points_data )
         self.createPolydata( **args )
